requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
requests==2.31.0
aiofiles==23.2.1
